Updated GSI to the newest version as well as the build scripts.

+++ sorc/rtma_gsi.fd/regression/regression_test.sh
@@ -8,10 +8,6 @@ exp1=$1
 exp2=$3
 exp3=$2
 
-#exp1=$global_T62_updat_exp1
-#exp2=$global_T62_contrl_exp1
-#exp3=$global_T62_updat_exp2
-
 input=$5
 
 #input=tmp62
@@ -37,7 +33,6 @@ cd $tmpdir
 
 # Other required constants for regression testing
 maxtime=1200
-maxmem=${maxmem:-3400000} # set in regression_param
 
 # Copy stdout and fort.220 files 
 # from $savdir to $tmpdir
@@ -49,6 +44,10 @@ for exp in $list; do
    $ncp $savdir/$exp/siginc ./siginc.$exp
    $ncp $savdir/$exp/wrf_inout ./wrf_inout.$exp
    $ncp $savdir/$exp/wrf_inout06 ./wrf_inout06.$exp
+   $ncp $savdir/$exp/siginc.nc ./siginc.nc.$exp
+   $ncp $savdir/$exp/fv3_dynvars ./fv3_dynvars.$exp
+   $ncp $savdir/$exp/fv3_sfcdata ./fv3_sfcdata.$exp
+   $ncp $savdir/$exp/fv3_tracer ./fv3_tracer.$exp
 done
 
 # Grep out penalty/gradient information, run time, and maximum resident memory from stdout file
@@ -71,9 +70,6 @@ diff penalty.$exp1.txt penalty.$exp3.txt > penalty.${exp1}-${exp3}.txt
 # Give location of additional output files for scalability testing
 exp1_scale=$2
 exp2_scale=$4
-
-#exp1_scale=$global_T62_updat_exp2
-#exp2_scale=$global_T62_contrl_exp2
 
 # Copy stdout for additional scalability testing
 list="$exp1_scale $exp2_scale"
@@ -185,24 +181,6 @@ fi
 
    } >> $output
 
-   # Next, maximum residence set size (both harware limitation and percent difference)
-   # First, hardware limitation
-
-   {
-
-     if [[ $(awk '{ print $8 }' memory.$exp1.txt) -gt $maxmem ]]; then
-       echo 'The memory for '$exp1' is '$(awk '{ print $8 }' memory.$exp1.txt)' KBs.  This has exceeded maximum allowable hardware memory limit of '$maxmem' KBs,'
-       echo 'resulting in Failure maxmem of the regression test.'
-       echo
-       failed_test=1
-     else
-       echo 'The memory for '$exp1' is '$(awk '{ print $8 }' memory.$exp1.txt)' KBs and is within the maximum allowable hardware memory limit of '$maxmem' KBs,'
-       echo 'continuing with regression test.'
-       echo
-     fi
-
-   } >> $output
-
    # Next, maximum residence set size
 
    {
@@ -226,18 +204,18 @@ fi
 
 if [[ $(grep -c 'cost,grad,step' penalty.${exp1}-${exp2}.txt) = 0 ]]; then
    if [[ $(grep -c 'congrad::evaljgrad: grepcost' penalty.${exp1}-${exp2}.txt) = 0 ]]; then
-      echo 'The results between the two runs ('${exp1}' and '${exp2}') are reproducible.'
+      echo 'The results (penalty) between the two runs ('${exp1}' and '${exp2}') are reproducible.'
 #      echo 'since the corresponding penalties and gradients are identical with '$(grep -c 'cost,grad,step' penalty.${exp1}-${exp2}.txt)' lines different.'
       echo
    else
-      echo 'The results between the two runs are nonreproducible,'
+      echo 'The results (penalty) between the two runs are nonreproducible,'
       echo 'thus the regression test has Failed on cost for '${exp1}' and '${exp2}' analyses.'
 #     echo 'thus the regression test has failed for '${exp1}' and '${exp2}' analyses with '$(grep -c 'cost,grad,step' penalty.${exp1}-${exp2}.txt)' lines different.'
       echo
       failed_test=1
    fi
 else
-   echo 'The results between the two runs are nonreproducible,'
+   echo 'The results (penalty) between the two runs are nonreproducible,'
    echo 'thus the regression test has Failed on cost for '${exp1}' and '${exp2}' analyses.'
    echo
 fi
@@ -317,7 +295,7 @@ else
 fi
 
 } >> $output
-   else
+   elif [[ -f siganl.${exp1} ]]; then
 {
 
 if cmp -s siganl.${exp1} siganl.${exp2} 
@@ -333,7 +311,57 @@ else
 fi
 
 } >> $output
+   elif [[ -f siginc.nc.${exp1} ]] ; then
+{
+ncdump siginc.nc.${exp1} > siginc.nc.${exp1}.out
+ncdump siginc.nc.${exp2} > siginc.nc.${exp2}.out
+if diff -s siginc.nc.${exp1}.out siginc.nc.${exp2}.out
+then
+   echo 'The results between the two runs ('${exp1}' and '${exp2}') are reproducible'
+   echo 'since the corresponding results are identical.'
+   echo
+else
+   echo 'The results between the two runs ('${exp1}' and '${exp2}') are not reproducible'
+   echo 'Thus, the case has Failed siganl of the regression tests.'
+   echo
+   failed_test=1
+fi
+} >> $output
    fi
+
+   elif [[ `expr substr $exp1 1 4` = "rrfs" ]] || [[ `expr substr $exp1 1 4` = "hafs" ]]; then
+{
+     fv3_failed_test=0
+     if cmp -s fv3_dynvars.${exp1} fv3_dynvars.${exp2}
+     then
+       echo 'The fv3_dynvars are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if cmp -s fv3_sfcdata.${exp1} fv3_sfcdata.${exp2}
+     then
+       echo 'The fv3_sfcdata are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if cmp -s fv3_tracer.${exp1} fv3_tracer.${exp2}
+     then
+       echo 'The fv3_tracer are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if [[ $fv3_failed_test -eq 0 ]]
+     then
+        echo 'The results between the two runs ('${exp1}' and '${exp2}') are reproducible'
+        echo 'since the corresponding results are identical.'
+        echo
+     else
+        echo 'The results between the two runs ('${exp1}' and '${exp2}') are not reproducible'
+        echo 'Thus, the case has Failed siganl of the regression tests.'
+        echo
+        failed_test=1
+     fi
+} >> $output
 fi
 
 # Next, reproducibility between exp1 and exp3
@@ -342,18 +370,18 @@ fi
 
 if [[ $(grep -c 'cost,grad,step' penalty.${exp1}-${exp3}.txt) = 0 ]]; then
    if [[ $(grep -c 'congrad::evaljgrad: grepcost' penalty.${exp1}-${exp3}.txt) = 0 ]]; then
-      echo 'The results between the two runs ('${exp1}' and '${exp3}') are reproducible'
+      echo 'The results (penalty) between the two runs ('${exp1}' and '${exp3}') are reproducible'
 #     echo 'since the corresponding penalties and gradients are identical with '$(grep -c 'cost,grad,step' penalty.${exp1}-${exp3}.txt)' lines different.'
       echo
    else
-      echo 'The results between the two runs are nonreproducible,'
+      echo 'The results (penalty) between the two runs are nonreproducible,'
       echo 'thus the regression test has Failed cost for '${exp1}' and '${exp3}' analyses.'
 #     echo 'thus the regression test has failed for '${exp1}' and '${exp3}' analyses with '$(grep -c 'cost,grad,step' penalty.${exp1}-${exp3}.txt)' lines different.'
       echo
       failed_test=1
    fi
 else
-   echo 'The results between the two runs are nonreproducible,'
+   echo 'The results (penalty) between the two runs are nonreproducible,'
    echo 'thus the regression test has Failed cost for '${exp1}' and '${exp3}' analyses.'
    echo
 fi
@@ -433,7 +461,8 @@ elif [[ `expr substr $exp1 1 6` = "global" ]]; then
       fi
 
 } >> $output
-   else
+
+   elif [[ -f siganl.${exp1} ]]; then
 
 {
 
@@ -450,34 +479,83 @@ elif [[ `expr substr $exp1 1 6` = "global" ]]; then
       fi
 
 } >> $output
+
+   elif [[ -f siginc.nc.${exp1} ]]; then
+
+{
+      ncdump siginc.nc.${exp1} > siginc.nc.${exp1}.out
+      ncdump siginc.nc.${exp3} > siginc.nc.${exp3}.out
+
+      if diff -s siginc.nc.${exp1}.out siginc.nc.${exp3}.out
+      then
+         echo 'The results between the two runs ('${exp1}' and '${exp3}') are reproducible'
+         echo 'since the corresponding results are identical.'
+         echo
+      else
+         echo 'The results between the two runs ('${exp1}' and '${exp3}') are not reproducible'
+         echo 'Thus, the case has Failed siganl of the regression tests.'
+         echo
+         failed_test=1
+      fi
+} >> $output
+
    fi
+
+elif [[ `expr substr $exp1 1 4` = "rrfs" ]] || [[ `expr substr $exp1 1 4` = "hafs" ]]; then
+{
+     fv3_failed_test=0
+     if cmp -s fv3_dynvars.${exp1} fv3_dynvars.${exp3}
+     then
+       echo 'The fv3_dynvars are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if cmp -s fv3_sfcdata.${exp1} fv3_sfcdata.${exp3}
+     then
+       echo 'The fv3_sfcdata are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if cmp -s fv3_tracer.${exp1} fv3_tracer.${exp3}
+     then
+       echo 'The fv3_tracer are reproducible'
+     else
+       fv3_failed_test=1
+     fi
+     if [[ $fv3_failed_test -eq 0 ]]
+     then
+        echo 'The results between the two runs ('${exp1}' and '${exp3}') are reproducible'
+        echo 'since the corresponding results are identical.'
+        echo
+     else
+        echo 'The results between the two runs ('${exp1}' and '${exp3}') are not reproducible'
+        echo 'Thus, the case has Failed siganl of the regression tests.'
+        echo
+        failed_test=1
+     fi
+} >> $output
+
 fi
-
-   # Finally, scalability
-
-   {
-
-   timelogic=$( echo "$scale1thresh >= $scale2" | bc )
-   if [[ "$timelogic" = 1 ]]; then
-      echo 'The case has passed the scalability regression test.'
-      echo 'The slope for the update ('$scale1thresh' seconds per node) is greater than or equal to that for the control ('$scale2' seconds per node).'
-   else
-      echo 'The case has Failed the scalability test.'
-      echo 'The slope for the update ('$scale1thresh' seconds per node) is less than that for the control ('$scale2' seconds per node).'
-   fi
-
-   } >> $output
 
 # Copy select results to $savdir
 mkdir -p $vfydir
 
 $ncp $output                        $vfydir/
 
+# Final check for any failed tests
+count=$(grep -i "fail" $output |wc -l)
+if [ $count -gt 0 ]; then
+    (( failed_test = $failed_test + $count ))
+fi
+
+# Remove job log files is no failures detected
 cd $scripts
-rm -f ${exp1}.out
-rm -f ${exp2}.out
-rm -f ${exp3}.out
-rm -f ${exp2_scale}.out
+if [ $count -eq 0 ]; then
+    rm -f ${exp1}.out
+    rm -f ${exp2}.out
+    rm -f ${exp3}.out
+    rm -f ${exp2_scale}.out
+fi
 
 if [[ "$clean" = ".true." ]]; then
    rm -rf $savdir
